refactor: Extract ViewerWrapper from Viewer main file
